+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
+- install mex-common and mex-artificial from pypi instead of github
+
 ### Deprecated
 
 ### Removed