+++ CHANGELOG.md
@@ -11,9 +11,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
-- bump mex-common dependency to 0.62.0
-- install mex-common and mex-artificial from pypi instead of github
-
 ### Deprecated
 
 ### Removed
@@ -21,6 +18,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 ### Security
+
+## [0.35.0] - 2025-06-17
+
+### Changes
+
+- bump mex-common dependency to 0.62.0
+- install mex-common and mex-artificial from pypi instead of github
 
 ## [0.34.5] - 2025-06-16